moving ili/inference/loaders.py to ili/dataloaders/loaders.py and making associated compatibility changes
