+++ examples/configs/sample_quijoteTPCF_mdn_config.yaml
@@ -1,5 +1,5 @@
 loader:
-  module: 'ili.inference.loaders'
+  module: 'ili.dataloaders'
   class: 'SummarizerDatasetLoader'
   args:
     stage: 'train'
